1227 운용사 팝업

+++ src/main/java/kr/co/kfs/assetedu/repository/Com02CodeRepository.java
@@ -7,5 +7,5 @@ import kr.co.kfs.assetedu.model.QueryAttr;
 
 public interface Com02CodeRepository {
 	List<Com02Code> selectList (QueryAttr queryAttr);
-	Integer selectCount(QueryAttr queryAttr);
+	Long selectCount(QueryAttr queryAttr);
 }
